pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tita-benchmarks"
version = "0.1.0"
description = "Benchmarks and API-based evaluation tools for Vision-Language models (TITA project)"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "TITA Contributors" }
]
keywords = ["vision-language", "benchmark", "LLM", "VLM", "evaluation"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "openai>=1.40.0",
  "pillow>=10.0.0",
  "tqdm>=4.65.0",
  "numpy>=1.24.0",
  "pandas>=2.0.0",
  "matplotlib>=3.7.0",
  "seaborn>=0.12.0"
]

[project.scripts]
tita-bench = "tita_benchmarks.api.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["tita_benchmarks*"]
exclude = ["tests*"]


